Fix the bug that caused caching to fail
My first proper Haskell debugging experience!
The state monad *almost* makes it possible to
forget that you are programming in an immutable
langauge. The bug was that I was replacing
the board cache with the 'state' from earlier
in the computation. This was the discarding
all of the state updates that had happened from
the recursive call to scoreM